Some updates for Investigation, Study and Assay profiles
This PR removes several properties of Assay from the profile that seem to be copied from Study, as they do not correspond to any ISA properties. An exception is the creator/performer of an assay. I kept it in there, since it is also part of an ARC Assay. Furthermore, some Study and Investigation properties were changed from MUST to SHOULD.